# packages/contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: PTC Bale Inspection Platform API
  version: 1.2.1
  description: >-
    Local-first API for the approved PTC dashboard, PostgreSQL platform,
    durable edge delivery, protected evidence, reporting and operations.
    Simulator output is never actual AI or client-acceptance evidence.
servers:
  - url: /api
security:
  - sessionCookie: []
paths:
...
  /evidence/{evidenceId}/content:
    parameters:
      - { name: evidenceId, in: path, required: true, schema: { type: string, maxLength: 96 } }
      - { name: Range, in: header, required: false, schema: { type: string, examples: [bytes=0-1023] } }
    get:
      operationId: evidenceContent
      responses:
        '200': { description: Complete authorized evidence content }
        '206': { description: Authorized byte range for evidence-clip playback }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '416': { $ref: '#/components/responses/Error' }
...
  /ingest/evidence/{eventId}:
    parameters:
      - { $ref: '#/components/parameters/EventId' }
      - { name: X-Evidence-ID, in: header, required: true, schema: { type: string, maxLength: 96 } }
      - { name: X-Checksum-SHA256, in: header, required: true, schema: { type: string, pattern: '^[a-fA-F0-9]{64}$' } }
      - { name: X-Duration-MS, in: header, required: false, schema: { type: integer, minimum: 0, maximum: 600000 } }
    post:
      security:
        - serviceBearer: []
      operationId: uploadEvidence
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          video/mp4: { schema: { type: string, format: binary } }
      responses:
        '201': { description: Evidence atomically stored and finalized }
        '200': { description: Exact evidence replay acknowledged }
        '409': { $ref: '#/components/responses/Error' }
        '415': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
  /operations/evidence/reconcile:
    post:
      operationId: reconcileEvidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                action: { type: string, enum: [report, mark-missing, quarantine-orphans], default: report }
                dryRun: { type: boolean, default: true }
      responses:
        '200': { description: Dry-run or audited reconciliation result }
        '403': { $ref: '#/components/responses/Error' }
  /operations/evidence/retention:
    post:
      operationId: evidenceRetention
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [olderThanDays]
              properties:
                olderThanDays: { type: integer, minimum: 1, maximum: 3650 }
                dryRun: { type: boolean, default: true }
      responses:
        '200': { description: Dry-run or audited deletion result; pending and quarantined evidence is protected }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ptc_session
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
  responses:
    Error:
      description: Stable API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        status: { type: integer }
        correlationId: { type: string }
        details: { type: object, additionalProperties: true }